Automatically go to the first match, if no matches are visible.
With this change, if no matches are visible anywhere an idle timer is
scheduled to search the buffer for a match, step by step, until on is
found. The search is done from the original point to the end, then to
the beginning to the original point.

This guarantees that there will always be at least one example of string
to be replaced that's shown, if there is one in the buffer.