Fix tests using test-visual-replace-overlay-to-properties.
With the introduction of a highlight for the scope, there are now
positions with two overlays. Before this change,
test-visual-replace-overlay-to-properties would be confused by it, as it
wasn't aware of overlay priorities.

This change replaces test-visual-replace-overlay-to-properties with a
new implementation that is based on visual-replace-test-window-content.
This also somewhat unifies the snapshots used in older tests with the
snapshots used in newer tests.