Navigate through matches with up and down.
With this change, pressing up and down moves the point to the next or
previous match. In the default keybinding C-p C-n remain available to
navigate through history.

This change also updates the documentation and readme to mention
scrolling the window and moving the point to keep at least one match
visible.

This partially addresses issue #6