Highlight the region using visual-replace-region, when active.
Before this change, when starting a replacement with the region active,
the region would remain highlighted for the duration of the command,
even when the scope for the replacement was the whole buffer.

With this change, the region is highlighted only when it is used as the
scope for the replacement.

This change also make the face visual-replace-region inherit from region
instead of secondary-selection.