Improve the way rectangular regions are highlighted.
When a rectangular region is highlighted, the right boundaries might not
fall on any actual text in the buffer.

Before this change, only the text that was part of the rectangular
region was highlighted, which might look strange.

With this change, when mark-rectangle-mode is used, visual-replace makes
sure to add as 'before-string or 'after-string any spaces needed to
display a series of overlays of the same width, even if text is missing.
This looks a little better.

It unfortunately won't work with any other rectangular selection mode.