Fix replacement in rectangles and other non-contiguous regions.
Before this change, when multiple ranges were passed to visual-replace,
it would execute the replacements one after the other. This didn't work
with things like \#, which increment a counter every time a replacement
was executed. This also cause the final message displayed by the
perform-replace to under-count the number of replacements.

With this change, non-contiguous ranges are handled by perform-replace
as non-contiguous regions normally are.