feat!: visual-replace-yank now yanks symbol, not words.
Before this change, a visual-replace-yank in a visual-replace prompt
would copy the word at point in the buffer. This was inconvenient when
trying to replace a variable in a language that uses word separators in
symbol, as it required calling visual-replace-yank multiple times.

Copying an entire symbol is more useful.

With this change, the "yank" copies the symbol at point into the prompt,
which is generally faster.