refactor: Simplify match overlay setup.
Before this change, overlays were built very differently for match with
an empty replacement than for match with a non-empty replacement. This
change makes them work the same way and cleans up
visual-replace--overlay.

This refactoring affected tests, which were unnecessarily dependent on
finding the original match in a "display" property, which needed to be
updated. Visible behavior should stay the same, however.