feat: Highlight match at point.
Before this change, when navigating through matches with up and down, it
was difficult to see what the current match was. While it is highlighted
by a cursor, the cursor of unselected windows by default gets hollow and
harder to see.

To fix that, this change introduces highlighted version of the match
faces, used for the match that contains the point.

Issue #7