fix: Only rebind this-command-keys if bound to 'visual-replace.
Before this change, visual-replace bound (this-command-keys-vector),
that is the sequence of keys that triggered visual-replace, to the
secondary mode map.

The idea was to reuse a key binding that was not useful, because it
would call visual-replace recursively.

This backfires in the case where visual-replace is called from a
non-global map, because then the key binding is *not* free and binding
it causes very strange behavior.

After this change, the binding only happens if the sequence of key that
triggered visual-replace would call visual-replace, if called in the
minibuffer, which should cause less issues.

Additionally, this change moves local-set-key after turning on
visual-replace-minibuffer-mode, so the local map is
visual-replace-mode-map already.

fixes #11