docs: Push for configuring keybinding for secondary-mode map.
By default, secondary-mode-map is bound to the key that's used to call
visual-replace. That works in some common cases, but easily causes
issues and is probably something that should be removed in future
versions.

It seems a better idea overall to push to setting up the binding for
visual-replace-secondary-mode-map explicitly in the example
configuration, is this change does.

Issue #11