feat: visual-replace-thing-at-point go to replacement area.
Before this change, when visual-replace-thing-at-point was called, it
opened Visual Replace with the prompt at the beginning of the match
area.

However, in most cases, when calling this function, we just want to
match the thing that was at point.

With this change, the point goes directly at the beginning of the
replacement area, saving a TAB in most cases.

See discussion in issue #9