feat: TAB moves point at end by default.
Before this change, when TAB had no marker history, it would go to the
beginning of the area.

With this change, pressing TAB moves the point at the end of the area,
which is more convenient, as we often want to extend the search or
replacement rather than rewrite it.

Fixes #11