refactor: Rewrite visual-replace--update-preview.
This change reworks the way the preview and first matches are computed.

With this change, the set of searches to run, including any search for
first match, are collected into a work queue, which is then executed by
visual-replace--run-idle-search.

visual-replace--run-idle-search runs the first work queue element
directly and schedules the rest, if any, to be run in an idle timer.

This change is meant to make it easier to introduce logic for counting
all matches within a buffer for issue #8 in an idle timer without
duplicating work done anyways for building the preview or looking for
the first match, in a future change.