feat: visual-replace-selected go to replacement area.
Before this change, when visual-replace-selected was called, it opened
Visual Replace with the prompt in match area.

However, in most cases, when calling this function, we just want to
match the thing that was at point.

With this change, the point goes directly at the beginning of the
replacement area, saving a TAB in most cases.

See discussion in issue #9