refactor: Avoid always rebuilding overlays when updating preview.
Before this change, visual-replace--update-preview would just blindly
regularly run a search in the visible ranges and re-build the overlays.

With this change visual-replace--update-preview tries to be more clever,
and, if possible:
 - don't do anything if the buffer, the visual-replace-args and
   visible range haven't changed
 - if only the point has changed, only update the highlights of the
   existing overlays
 - if only the replacement has changed, only update the replacements
   of the existing overlays
And only re-build the overlays and possibly search for the first match
if all that fails.

The goal of this change is less an optimization and more a
re-organization to make it easier to extend and deal, in the future,
with a situation where overlays would be available for all matches in a
buffer.