feat: Display total match count.
This change introduces the option visual-replace-display-total, which,
if enabled, instructs visual-replace to create overlays for the whole
buffer in an idle timer and, once the buffer has been fully searched,
display the total number of matches in the prompt.

This requires running in two modes, in one mode, overlays are only
available for the visible ranges and built as necessary. In another
mode, once overlays have been created for the whole buffer, they just
need to updated as necessary.

Issue #8