feat: Highlight match at point when there's no replacement.
Before this change, matches without replacement strings were always
displayed the same, no matter where the point was, and even though
navigating up and down the set of matches worked.

With this change, matches without replacement containing the point get
the new face visual-replace-match-highlight instead of
visual-replace-match, so we can better see which match is currently
active for replacement and movement.