feat: Deal with window configuration changes during session.
Before this change, Visual Replace would store the calling window and
use that from then on, assuming that it would be always available and
always show the right buffer.

With this change, Visual Replace checks whether the calling window it
stored is still usable and if not looks for another window showing
buffer - displaying the buffer if necessary.