feat: Define fields for the search and replacement strings.
This change defines two separate fields in the minibuffer when the
separator is shows: one for the search string and one for the
replacement. This way, standard Emacs commands such as kill,
beginning-of-line and end-of-line behave reasonably.

This does noticeably change the behavior of C-a and C-e, which now go to
the beginning and end of each field separately. TAB is now the best way
to switch between fields.

This change also removes visual-replace-kill and
visual-replace-kill-whole-line. They're not useful anymore, since kill
and kill-whole-line behave as they should in a field.