feat: When the point is moved during preview, stay there.
Before this change, Visual Replace would always go back to the position
it was originally called from. This could be jarring in the cases where
the point was moved during preview to find the first match, as the
effect of replacement wouldn't be visible.

With this change, the point stays where it was moved to during preview.
A mark is pushed to go back to the original position if necessary.

This change introduces a noticeable change in the way Visual Replace
behaves. To go back to original behavior, set the customization option
visual-replace-keep-initial-position to t.