feat: Limit the size of ranges to be searched.
This change introduce a byte limit for computing the total as well as
looking for the first match. This avoid slowing down Emacs too much when
running visual-replace on a very large buffer.

When the limit is hit, preview still works, as it only requires
searching the visible portions of the buffer, but computing the total
won't and jumping to the first match might miss matches that are too far
away.