feat: Automatically go to the closest match during preview.
Before this change, Visual Replace would only move the point if there
was no match currently displayed in any window. The point would only
move when using <up> or <down>. This behavior was unexpected, as isearch
always moves the point to a match.

With this change, Visual Replace moves to the visible match closest to
the point, during preview. This way, the point start by being on a match
without having to do anything.