refactoring: Merge show-first-match and goto-closest-match.
Before this change, show-first-match was responsible for moving the
window to display the first match and goto-closest-match for going to
the closest visible match.

This change merges the two into one function that goes to the closest
match it is given, then moves the window if necessary, without bothering
to check whether it is in the visible range.