test: Update test-visual-replace-read-apply-undo-everything-then-redo
This change rewrites
test-visual-replace-read-apply-undo-everything-then-redo to use turtles.

It also switches to using M-x <command name> instead of using keyboard
shortcuts, as it makes it easier to see what is executed.

Using M-x <command> instead of calling the command directly guarantees
that everything that Emacs executes around a command is included - in
this test, that means undo boundaries are added normally.