test: Fix test for highlight-scope.
Before this change, tests for highlight-scope treated this face as if it
had :extend nil and would highlight it line-by-line.

With this change and the fix in turtles brought in this commit,
multi-line regions with this face are treated as one region.

This also fixes tests for highlight-scope so they run under Emacs 26 and
29.