fix: Disable showing keymap help automatically; it swallows keys.
The feature for automatically showing help, introduced in fe65cc0,
causes the keystroke that closes the help to be swallowed when the help
is shown.

This change disables the feature by default and adds a warning not to
enable. This gives some time to either fixing the bug or cleanly
removing it.