fix: Don't let which-key swallow keystrokes when shown automatically.
Before this change, when visual-replace-show-mode-map-help was non-nil
and which-key was shown by visual-replace, the keystroke that turns off
the help was swallowed, by which-key, it turns out.

This changes fixes the problem by telling which-key-show-keymap not to
read key to allow paging.

This change also re-enables visual-replace-show-mode-map-help, now that
the issue as been fixed.