feat: Add dropdown menu on the prompt to edit options.
This changes defines a menu that pops up when the prompt is clicked.
This menu presents the current search and replace options and lets the
user modify them.

This also doubles as a way of reminding the user which keys they can use
to change options.

This change also adds some commands that only exist to be bound to a
menu; as just "toggle scope" isn't very descriptive.

Note that the menu does not appear in the main menu bar, because it
seems that menu in the menu bar is not systematically updated, at
least on macos, and the result is more confusing than helpful.