docs: Reword comparison section.
The text of the comparison section made it sound like only visual
replace offers users the capability of editing both the search string
and the replacement string at a time. That's not true: query-regexp at
least lets you do that, but only when reusing replacement from history.
What's special about Visual replace is that this mode is the normal
case.