fix: Avoid infinite loop when applying one recursive replacement
Before this change, perform-replace would enter an infinite loop when
trying to apply a single replacement(1) that contains the search string,
such as when replacing "foo" with "foor".

This changes avoids triggering that behavior in perform-replace and also
attempts to clean up the preview behavior and cursor position when doing
single replacements.

fixes #17

(1) actually a replacement within a very small range that corresponds to
a single match